- Moved list_eq from test_multi to dataframe/utils testing area - Moved melt test from test_multi to test_reshape and imported list_eq - Added additional test for melt as DataFrame method
